+++ server/ee/libs/platform/platform-data-storage/platform-data-storage-db/platform-data-storage-db-remote-rest/build.gradle.kts
